+++ src/Physics/Point.hpp
@@ -12,17 +12,17 @@ namespace pong {
     Point &operator=(const Point &);
     ~Point() = default;
 
-    Point(int, int);
+    Point(float, float);
     Point operator+(const Point&) const;
     Point operator-(const Point&) const;
     Point &operator+=(const Point&);
     Point &operator-=(const Point&);
 
-    int distanceTo(const Point&) const;
+    float distanceTo(const Point&) const;
     std::string to_string() const;
 
-    int x;
-    int y;
+    float x;
+    float y;
   };
 }
